[update] : Updated front end code, based on designs
