+++ README.md
@@ -1,2 +1,16 @@
 # Udyamitra
 Udyamitra is an intelligent AI-powered assistant designed to empower India's Micro, Small, and Medium Enterprises (MSMEs) and manufacturing sectors by making government schemes, incentives, and compliance processes more accessible and actionable.
+
+### Commands
+
+#### Frontend
+- Go to the frontend folder : cd frontend
+- Run the frontend files : npm run dev
+
+#### Backend
+- From the root directory of the project, run command:
+- uvicorn Servers.backend:app --reload
+
+#### MCP Servers
+- From the root directory of the project, run command:
+- uvicorn Servers.main:server --reload --port 10000
