+++ frontend/.gitignore
@@ -11,9 +11,6 @@ node_modules
 dist
 dist-ssr
 *.local
-.env
-.env.*
-!.env.example
 
 # Editor directories and files
 .vscode/*
