+++ frontend/index.html
@@ -2,9 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Udyamitra</title>
-    <meta name="description" content="Udyamitra is an intelligent AI-powered assistant designed to empower India's Micro, Small, and Medium Enterprises (MSMEs) and manufacturing sectors by making government schemes, incentives, and compliance processes more accessible and actionable." />
+    <title>Vite + React</title>
   </head>
   <body>
     <div id="root"></div>
