+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tech-trends-chatbot-frontend",
+  "name": "chatbot",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -10,26 +10,23 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "eventsource-parser": "^3.0.2",
-    "immer": "^10.1.1",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
+    "@tailwindcss/vite": "^4.1.12",
+    "lucide-react": "^0.542.0",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
     "react-markdown": "^10.1.0",
+    "tailwindcss": "^4.1.12",
     "use-immer": "^0.11.0"
   },
   "devDependencies": {
-    "@eslint/js": "^9.28.0",
-    "@tailwindcss/postcss": "^4.1.8",
-    "@types/react": "^19.1.6",
-    "@types/react-dom": "^19.1.5",
-    "@vitejs/plugin-react": "^4.5.0",
-    "eslint": "^9.28.0",
-    "eslint-plugin-react": "^7.37.5",
+    "@eslint/js": "^9.33.0",
+    "@types/react": "^19.1.10",
+    "@types/react-dom": "^19.1.7",
+    "@vitejs/plugin-react": "^5.0.0",
+    "eslint": "^9.33.0",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
-    "globals": "^16.2.0",
-    "postcss": "^8.5.4",
-    "tailwindcss": "^4.1.8",
-    "vite": "^6.3.5"
+    "globals": "^16.3.0",
+    "vite": "^7.1.2"
   }
 }
