+++ frontend/src/index.css
@@ -1,110 +1,1 @@
-@import url('https://fonts.googleapis.com/css2?family=Urbanist:wght@100..900&display=swap') layer(base);
-@import url('https://fonts.googleapis.com/css2?family=Open+Sans:wght@300;400;600;700&display=swap') layer(base);
-@import 'tailwindcss';
-
-@theme {
-  --font-sans: Open Sans, sans-serif;
-  --font-urbanist: Urbanist, sans-serif;
-
-  /* Primary Color Scheme */
-  --color-background: #1E1E1E;
-  --color-user-message: #2E8B57;      /* Sea Green */
-  --color-assistant-message: #452C63; /* Dark Purple */
-  --color-main-text: #E5E5E5;
-  --color-muted-text: #A0A0A0;
-
-  /* Optional Accents */
-  --color-accent: #F3C3B1;
-  --color-accent-light: #92B3CA;
-  --color-error: #D04552;
-
-  /* Spinner animation */
-  --animate-spinner: spinner 1.5s linear infinite;
-  --animate-spinner-delayed: spinner 1.5s linear infinite 0.75s;
-
-  @keyframes spinner {
-    0% {
-      transform: scale(0);
-      opacity: 1;
-    }
-    100% {
-      transform: scale(1);
-      opacity: 0;
-    }
-  }
-}
-
-@utility container {
-  margin-inline: auto;
-  padding-inline: 2rem;
-  @media (width >= --theme(--breakpoint-sm)) {
-    max-width: none;
-  }
-  @media (width >= 1400px) {
-    max-width: 1400px;
-  }
-}
-
-@layer base {
-  * {
-    border-color: var(--color-gray-200, currentcolor);
-  }
-
-  body,
-  html,
-  #root {
-    @apply font-sans h-full;
-    background-color: var(--color-background);
-    color: var(--color-main-text);
-  }
-
-  ::-webkit-scrollbar {
-    @apply h-4 w-2;
-  }
-
-  ::-webkit-scrollbar-track {
-    @apply bg-transparent rounded-sm;
-  }
-
-  ::-webkit-scrollbar-thumb {
-    background-color: var(--color-accent-light);
-    border: 1px solid white;
-    border-radius: 0.25rem;
-  }
-}
-
-@utility spinner-child {
-  @apply absolute -top-1 left-0 h-full w-full content-[''] rounded-full;
-  background-color: var(--color-accent);
-  opacity: 0;
-}
-
-@utility markdown-container {
-  & p {
-    @apply mb-5;
-  }
-
-  & h1 {
-    @apply font-bold mb-5;
-  }
-
-  & li {
-    @apply mb-3;
-  }
-
-  & li:last-child {
-    @apply mb-0;
-  }
-
-  & ol {
-    @apply list-decimal mx-0 mt-0 mb-5 pl-8;
-  }
-
-  & ul {
-    @apply list-disc mx-0 mt-0 mb-5 pl-8;
-  }
-
-  & > :last-child {
-    @apply mb-0;
-  }
-}
+@import "tailwindcss";
