+++ frontend/vite.config.js
@@ -1,19 +1,9 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-import path from 'path';
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+import tailwindcss from '@tailwindcss/vite'
 
-// https://vitejs.dev/config/
+
+// https://vite.dev/config/
 export default defineConfig({
-  server: {
-    port: 3000,
-  },
-  preview: {
-    port: 3000,
-  },
-  resolve: {
-    alias: {
-      '@': path.resolve(__dirname, './src'),
-    },
-  },
-  plugins: [react()],
+  plugins: [react(), tailwindcss()],
 })
